Rename the miner file
