+++ miner/dynamic_attributes.py
